requirements.txt
cityscapesscripts==2.2.0
loguru==0.6.0
Pillow
rich
